Отработка замечаний наставника на созвоне 06/02/24. Вынесла type TRestSchema в отдельный файл. Исправила импорты. default-ные значения у параметров convict в rest.schema пришлось оставить. Убрала вывод логов в файл. Вынесла формирование контейнера зависимостей в отдельный модуль container в rest.
