+++ src/main.rest.ts
@@ -1,16 +1,9 @@
 import 'reflect-metadata';
-import { Container } from 'inversify';
-import { Logger, PinoLogger } from './libs/logger/index.js';
 import { RestApplication } from './rest/index.js';
-import { RestConfig, Config, TRestSchema } from './libs/config/index.js';
 import { Component } from './types/index.js';
+import { container } from './rest/container.js';
 
 async function bootstrap() {
-  const container = new Container();
-  container.bind<RestApplication>(Component.RestApplication).to(RestApplication).inSingletonScope();
-  container.bind<Logger>(Component.Logger).to(PinoLogger).inSingletonScope();
-  container.bind<Config<TRestSchema>>(Component.Config).to(RestConfig).inSingletonScope();
-
   const application = container.get<RestApplication>(Component.RestApplication);
   await application.init();
 }
