+++ src/types/rest-schema-type.ts
@@ -1,0 +1,5 @@
+export type TRestSchema = {
+  PORT: number;
+  SALT: string;
+  DB_HOST: string;
+}
